pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ca-language-model"
version = "0.1.0"
description = "Tiny experiments comparing a transformer baseline to a neural cellular automata language model."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["torch", "pillow"]
